requirements.txt
Django==2.2.2
djangorestframework==3.9.4
django-cors-headers==3.0.2
django-filter==2.1.0
Unipath==1.1
colorlog==4.0.2
ConcurrentLogHandler==0.9.1
PyMySQL==0.9.3
Pillow==6.0.0
celery==4.3.0
requests==2.20.0
django-phonenumber-field==3.0.1
phonenumbers==8.10.13